% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_catalogs_from_vcf.R
\name{generate_catalogs_from_mutations}
\alias{generate_catalogs_from_mutations}
\title{Generate catalogs from VCF data}
\usage{
generate_catalogs_from_mutations(
  Indel_VCF = NULL,
  SNV_VCF = NULL,
  genome.v = "hg38",
  sample_name = NULL
)
}
\arguments{
\item{Indel_VCF:}{Vector of paths or list of data frames. If it is a vector of paths the file will be rearranged by the function, otherwise it has to be a list of dataframe having the following structure: "Sample", "chr", "position", "REF", "ALT"}

\item{SNV_VCF:}{Vector of paths or list of data frames.}

\item{genome.v:}{Either "hg19" or "hg38", which is the default option.}

\item{sample_name:}{Vector of sample names, it has to be in the same order of Indel VCF vector or list.}
}
\value{
Returns a list containing the Indel Catalos, total InD and/or SNV catalogs, total SNV
}
\description{
This function returns the catalogs and the total number of mutations from VCF files. You can give just InDel VCF param or SNV VCF param.
}
