% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRRDetect.R
\name{prepare_PRRDetect}
\alias{prepare_PRRDetect}
\title{prepare_PRRDetect}
\usage{
prepare_PRRDetect(
  InDel_Catalogs,
  SNV_Catalogs,
  Indel_VCF_path,
  total_InDel,
  SNV_VCF_path,
  genome.v,
  nparallel,
  is.filtered_SNV = F,
  is.filtered_InDel = F,
  sample_name,
  organ,
  setseed = 42
)
}
\arguments{
\item{InDel_Catalogs:}{Indel Catalogs dataframe}

\item{SNV_Catalogs:}{SNV catalogs daframe}

\item{Indel_VCF_path:}{Vector of paths or list of data frames. If it is a vector of paths the file will be rearranged by the function, otherwise it has to be a list of dataframe having the following structure: "Sample", "chr", "position", "REF", "ALT"}

\item{total_InDel:}{Total number of indels, it can be defined previously}

\item{SNV_VCF_path:}{Vector of paths or list of data frames.}

\item{genome.v:}{Either "hg19" or "hg38", which is the default option.}

\item{nparallel:}{Number of threads for signature fit}

\item{is.filtered_SNV:}{if yes, SNV VCF get filtered. Default equal to False}

\item{is.filtered_InDel:}{if yes InDel VCF get filtered. Default equal to False}

\item{sample_name:}{Vector of sample names, it has to be in the same order of Indel VCF vector or list.}

\item{organ:}{Organ for SBS signature fit}

\item{setseed:}{Seed for reproducibility purposes}
}
\value{

}
\description{
Prepares data for PRRDetect input
}
