% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRRDetect.R
\name{PRRDetect}
\alias{PRRDetect}
\title{Compute PRRDetect probability and label}
\usage{
PRRDetect(InDel_fits, SNV_fits, total_SNV, total_InD, prediction_type = "prop")
}
\arguments{
\item{ind_fits:}{Indel Signatures Fits}

\item{snv_fits:}{SNV Signatures Fits}

\item{total_SNV:}{total number of SNVs of the sample}

\item{total_InD:}{total number of InDel mutation in the sample}
}
\value{
PRRDetect table
}
\description{
Compute PRRDetect probability and label
}
