% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_fit.R
\name{Signature_fit}
\alias{Signature_fit}
\title{Signature fit}
\usage{
Signature_fit(SNV_Catalog, InDel_Catalog, organ, nparallel, setseed)
}
\arguments{
\item{SNV_Catalog:}{SNV Catalog}

\item{InDel_Catalog:}{Indel Catalog}

\item{organ:}{Organ necessary for SNV signature fit}

\item{nparallel:}{Number of threads}

\item{setseed:}{seed for reproducibility}
}
\value{
List containing SNV fit and Indel fit
}
\description{
Signature fit
}
