% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRRDetect.R
\name{prepare_PRRDetect}
\alias{prepare_PRRDetect}
\title{prepare_PRRDetect}
\usage{
prepare_PRRDetect(
  InDel_Catalogs = NULL,
  SNV_Catalogs = NULL,
  Indel_VCF_path = NULL,
  total_InDel = NULL,
  SNV_VCF_path = NULL,
  genome.v = "hg38",
  nparallel = NULL,
  is.filtered_SNV = T,
  is.filtered_InDel = T,
  sample_name = NULL,
  organ = NULL,
  setseed = 42
)
}
\arguments{
\item{InDel_Catalogs:}{Indel Catalogs dataframe}

\item{SNV_Catalogs:}{SNV catalogs daframe}

\item{Indel_VCF_path:}{Vector of paths or list of data frames. If it is a vector of paths the file will be rearranged by the function, otherwise it has to be a list of dataframe having the following structure: "Sample", "chr", "position", "REF", "ALT"}

\item{total_InDel:}{Total number of indels, it can be defined previously}

\item{SNV_VCF_path:}{Vector of paths or list of data frames.}

\item{genome.v:}{Either "hg19" or "hg38", which is the default option.}

\item{nparallel:}{Number of threads for signature fit}

\item{is.filtered_SNV:}{if yes, SNV VCF get filtered. Default equal to False. SNV information gets filtered only if provided as a vector of paths.}

\item{is.filtered_InDel:}{if yes InDel VCF get filtered. Default equal to False. InDel information gets filtered only if provided as a vector of paths.}

\item{sample_name:}{Vector of sample names, it has to be in the same order of Indel VCF vector or list.}

\item{organ:}{Organ for SBS signature fit}

\item{setseed:}{Seed for reproducibility purposes}
}
\value{
A list containing the indel and snv fits, the total number of SNVs, total number of InDels. If InDel VCF paths are provided, the total number of InDels is the number of InDels in the VCF file, otherwise it is the sum of the mutations found in the catalog.
}
\description{
Prepares data for PRRDetect input
}
